refactor: move timedate builtin to time package
